+++ lib/src/migrators/module.dart
@@ -138,8 +138,8 @@ class _ModuleMigrationVisitor extends MigrationVisitor {
 
   /// Set of canonical URLs that have a `@use` rule in the current stylesheet.
   ///
-  /// This includes both `@use` rules migrated from `@import` rules and
-  /// additional `@use` rules in the sets below.
+  /// This includes rules migrated from `@import` rules, additional rules in the
+  /// sets below, and existing `@use` rules in the file prior to migration.
   Set<Uri> _usedUrls;
 
   /// Set of additional `@use` rules for built-in modules.
@@ -402,9 +402,17 @@ class _ModuleMigrationVisitor extends MigrationVisitor {
     var oldBeforeFirstImport = _beforeFirstImport;
     var oldAfterLastImport = _afterLastImport;
     var oldUseAllowed = _useAllowed;
-    _namespaces = _determineNamespaces(node.span.sourceUrl);
-    _forwardedUrls = {};
+    _namespaces = {
+      for (var rule in node.uses)
+        importCache
+                .canonicalize(rule.url,
+                    baseImporter: importer, baseUrl: node.span.sourceUrl)
+                ?.item2 ??
+            rule.url: rule.namespace
+    };
+    _determineNamespaces(node.span.sourceUrl, _namespaces);
     _usedUrls = {};
+    _forwardedUrls = {};
     _builtInUseRules = {};
     _additionalLoadPathUseRules = {};
     _additionalRelativeUseRules = {};
@@ -458,23 +466,20 @@ class _ModuleMigrationVisitor extends MigrationVisitor {
     addPatch(Patch.insert(insertionPoint, extras));
   }
 
-  /// Determines namespaces for all `@use` rules that the stylesheet at [url]
-  /// will contain after migration.
-  Map<Uri, String> _determineNamespaces(Uri url) {
-    var namespaces = <Uri, String>{};
+  /// Adds namespaces for the `@use` rules added to [url] by the migration.
+  ///
+  /// [namespaces] should contain any existing namespaces prior to migration
+  /// (i.e. from existing `@use` rules).
+  void _determineNamespaces(Uri url, Map<Uri, String> namespaces) {
     var sourcesByNamespace = <String, Set<ReferenceSource>>{};
     for (var reference in references.sources.keys) {
       if (reference.span.sourceUrl != url) continue;
       var source = references.sources[reference];
+      if (source is UseSource) continue;
+      if (namespaces.containsKey(source.url)) continue;
       var namespace = source.preferredNamespace;
       if (namespace == null) continue;
-
-      // Existing `@use` rules should always keep their namespaces.
-      if (source is UseSource) {
-        namespaces[source.url] = namespace;
-      } else {
-        sourcesByNamespace.putIfAbsent(namespace, () => {}).add(source);
-      }
+      sourcesByNamespace.putIfAbsent(namespace, () => {}).add(source);
     }
     // First assign namespaces to module URLs without conflicts.
     var conflictingNamespaces = <String, Set<ReferenceSource>>{};
@@ -490,7 +495,6 @@ class _ModuleMigrationVisitor extends MigrationVisitor {
     conflictingNamespaces.forEach((namespace, sources) {
       _resolveNamespaceConflict(namespace, sources, namespaces, url);
     });
-    return namespaces;
   }
 
   /// Resolves a conflict between a set of sources with the same default
@@ -1092,21 +1096,27 @@ class _ModuleMigrationVisitor extends MigrationVisitor {
     super.visitMixinRule(node);
   }
 
-  /// If [node] is for a built-in module, adds its URL to [_usedUrls] so we
-  /// don't add a duplicate one, but ignore other `@use` rules, as we'll assume
-  /// they've already been migrated.
+  /// Don't visit `@use` rules.
   ///
   /// The migrator will use the information from [references] to migrate
   /// references to members of these dependencies.
   void visitUseRule(UseRule node) {
-    if (node.url.scheme == 'sass') _usedUrls.add(node.url);
+    _usedUrls.add(importCache
+            .canonicalize(node.url, baseImporter: importer, baseUrl: currentUrl)
+            ?.item2 ??
+        node.url);
   }
 
   /// Similar to `@use` rules, don't visit `@forward` rules.
   ///
   /// The migrator will use the information from [references] to migrate
   /// references to members of these dependencies.
-  void visitForwardRule(ForwardRule node) {}
+  void visitForwardRule(ForwardRule node) {
+    _forwardedUrls.add(importCache
+            .canonicalize(node.url, baseImporter: importer, baseUrl: currentUrl)
+            ?.item2 ??
+        node.url);
+  }
 
   /// Adds a namespace to any variable that requires it.
   @override
@@ -1221,7 +1231,6 @@ class _ModuleMigrationVisitor extends MigrationVisitor {
     if (libraryUrls != null) {
       url = minBy(libraryUrls, (url) => url.pathSegments.length);
     }
-
     if (!_usedUrls.contains(url)) {
       // Add new `@use` rule for indirect dependency
       var tuple = _absoluteUrlToDependency(url);